refactored DataLakeScgedzker
