+++ streampipes-data-export/pom.xml
@@ -52,6 +52,11 @@
             <artifactId>streampipes-storage-management</artifactId>
             <version>0.99.0-SNAPSHOT</version>
         </dependency>
+            <dependency>
+            <groupId>org.apache.streampipes</groupId>
+            <artifactId>streampipes-data-explorer-management</artifactId>
+            <version>0.99.0-SNAPSHOT</version>
+        </dependency>
     </dependencies>
 
     <build>
